requirements.txt
asgiref==3.7.2
Django==4.1.12
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
phonenumbers==8.13.23
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pymongo==3.12.1
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.8.0
